@@ -80,15 +80,15 @@ public class Discography implements MusicServiceEventListener {
         }
     }
 
-    public void setStale(boolean value) {
+    void setCacheState(MemCache.ConsistencyState value) {
         synchronized (cache) {
-            cache.isStale = value;
+            cache.consistencyState = value;
         }
     }
 
-    public boolean isStale() {
+    MemCache.ConsistencyState getCacheState() {
         synchronized (cache) {
-            return cache.isStale;
+            return cache.consistencyState;
         }
     }
 
@@ -122,6 +122,33 @@ public class Discography implements MusicServiceEventListener {
             Song song = cache.songsById.get(songId);
             return song == null ? Song.EMPTY_SONG : song;
         }
+    }
+
+    @NonNull
+    public ArrayList<Song> getSongsFromIdsAndCleanupOrphans(@NonNull ArrayList<Long> songIds, @Nullable Consumer<ArrayList<Long>> orphanIdsCleaner) {
+        ArrayList<Long> orphanSongIds = new ArrayList<>();
+        ArrayList<Song> songs = new ArrayList<>();
+
+        synchronized (cache) {
+            for (Long id : songIds) {
+                Song song = cache.songsById.get(id);
+                if (song != null) {
+                    songs.add(song);
+                } else if (orphanIdsCleaner != null) {
+                    orphanSongIds.add(id);
+                }
+            }
+        }
+
+        MemCache.ConsistencyState cacheState = getCacheState();
+        // In the case where the Discog is being reset situation, the operation takes time,
+        // and while the cache is being filled up,
+        // correct and existing songs may be considered as orphan
+        // --> incorrectly cleaned from the auxiliary DBs (history, queue, etc)
+        if ((orphanIdsCleaner != null) && (cacheState != MemCache.ConsistencyState.RESETTING)) {
+            orphanIdsCleaner.accept(orphanSongIds);
+        }
+        return songs;
     }
 
     @NonNull
@@ -268,7 +295,7 @@ public class Discography implements MusicServiceEventListener {
     }
 
     public void triggerSyncWithMediaStore(boolean reset) {
-        if (isStale()) {
+        if (getCacheState() != MemCache.ConsistencyState.OK) {
             // Prevent reentrance - delay to later
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
                 final long DELAY = 500;
@@ -277,7 +304,7 @@ public class Discography implements MusicServiceEventListener {
                 mainActivityTaskQueue.postDelayed(() -> triggerSyncWithMediaStore(reset), TASK_QUEUE_COALESCENCE_TOKEN, DELAY);
             } // else: too bad, just drop the operation. It is unlikely we get there anyway
         } else {
-            (new SyncWithMediaStoreAsyncTask(mainActivity, this)).execute(reset);
+            (new SyncWithMediaStoreAsyncTask(mainActivity, this, reset)).execute();
         }
     }
 
@@ -412,13 +439,13 @@ public class Discography implements MusicServiceEventListener {
     }
 
     private void fetchAllSongs() {
-        setStale(true);
+        setCacheState(MemCache.ConsistencyState.REFRESHING);
 
         Collection<Song> songs = database.fetchAllSongs();
         for (Song song : songs) {
             addSong(song, true);
         }
 
-        setStale(false);
+        setCacheState(MemCache.ConsistencyState.OK);
     }
 }
