@@ -57,7 +57,8 @@ public class MediaStoreBridge {
 
     @NonNull
     private static Song getSongFromCursorImpl(@NonNull Cursor cursor) {
-        // TODO Consider ignoring most of these columns, we do extract them ourselves already
+        // Most of the time data imported from these columns are overriden by those extracted ourselves from ID3 tags
+        // However, in the case extraction fails (unsupported track format), they serve as fallback values
         final long id = cursor.getLong(0);
         final String title = cursor.getString(1);
         final int trackNumber = cursor.getInt(2);
@@ -71,7 +72,14 @@ public class MediaStoreBridge {
         final long artistId = cursor.getLong(10);
         final List<String> artistNames = MultiValuesTagUtil.split(cursor.getString(11));
 
-        return new Song(id, title, trackNumber, year, duration, data, dateAdded, dateModified, albumId, albumName, artistId, artistNames);
+        final Song song = new Song(id, title, trackNumber, year, duration, data, dateAdded, dateModified, albumId, albumName, artistId, artistNames);
+
+        // MediaStore compat: Split track number into disc + track number
+        // See documentation for MediaStore.Audio.AudioColumns.TRACK
+        song.discNumber = trackNumber / 1000;
+        song.trackNumber = trackNumber % 1000;
+
+        return song;
     }
 
     @Nullable
