@@ -23,8 +23,12 @@ import java.util.function.Function;
  */
 
 class MemCache {
-    // Indicate whether the cache can be considered as usable/reliable
-    boolean isStale = true;
+    enum ConsistencyState {
+        RESETTING,
+        REFRESHING,
+        OK
+    };
+    ConsistencyState consistencyState = ConsistencyState.REFRESHING;
 
     final Map<Long, Song> songsById = new HashMap<>();
 
