@@ -4,11 +4,11 @@ import android.os.Parcel;
 import android.os.Parcelable;
 import android.text.TextUtils;
 
-import com.poupa.vinylmusicplayer.util.ComparatorUtil;
 import com.poupa.vinylmusicplayer.util.MusicUtil;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.Objects;
 
 /**
@@ -60,9 +60,11 @@ public class Artist implements Parcelable {
     }
 
     public long getDateModified() {
+        if (albums.isEmpty()) {return Song.EMPTY_SONG.dateModified;}
+
         return Collections.max(
                 albums,
-                (a1, a2) -> ComparatorUtil.compareLongInts(a1.getDateModified(), a2.getDateModified())
+                Comparator.comparingLong(Album::getDateModified)
         ).getDateModified();
     }
 
