@@ -27,21 +27,22 @@ public class SongSortOrder {
             MultiValuesTagUtil.infoString(s1.artistNames),
             MultiValuesTagUtil.infoString(s2.artistNames));
     private static final Comparator<Song> _BY_ALBUM = (s1, s2) -> StringUtil.compareIgnoreAccent(s1.albumName, s2.albumName);
-    private static final Comparator<Song> _BY_YEAR = (s1, s2) -> s1.year - s2.year;
-    public static final Comparator<Song> BY_DATE_ADDED = (s1, s2) -> ComparatorUtil.compareLongInts(s1.dateAdded, s2.dateAdded);
+    private static final Comparator<Song> _BY_ALBUM_ID = Comparator.comparingLong(s -> s.albumId); // to combine with comparison by name to make the sorting deterministic in the case the names are identical
+    private static final Comparator<Song> _BY_YEAR = Comparator.comparingInt(s -> s.year);
+    public static final Comparator<Song> BY_DATE_ADDED = Comparator.comparingLong(s -> s.dateAdded);
     public static final Comparator<Song> BY_DATE_ADDED_DESC = ComparatorUtil.reverse(BY_DATE_ADDED);
-    private static final Comparator<Song> BY_DATE_MODIFIED = (s1, s2) -> ComparatorUtil.compareLongInts(s1.dateModified, s2.dateModified);
+    private static final Comparator<Song> BY_DATE_MODIFIED = Comparator.comparingLong(s -> s.dateModified);
     private static final Comparator<Song> BY_DATE_MODIFIED_DESC = ComparatorUtil.reverse(BY_DATE_MODIFIED);
-    public static final Comparator<Song> BY_DISC_TRACK = (s1, s2) -> (s1.discNumber != s2.discNumber)
+    private static final Comparator<Song> _BY_DISC_TRACK = (s1, s2) -> (s1.discNumber != s2.discNumber)
             ? (s1.discNumber - s2.discNumber)
             : (s1.trackNumber - s2.trackNumber);
-
-    private static final Comparator<Song> BY_TITLE = ComparatorUtil.chain(_BY_TITLE, _BY_ARTIST);
-    private static final Comparator<Song> BY_TITLE_DESC = ComparatorUtil.chain(ComparatorUtil.reverse(_BY_TITLE), _BY_ARTIST);
-    private static final Comparator<Song> BY_ARTIST = ComparatorUtil.chain(_BY_ARTIST, _BY_TITLE);
-    public static final Comparator<Song> BY_ALBUM = ComparatorUtil.chain(_BY_ALBUM, BY_DISC_TRACK);
-    private static final Comparator<Song> BY_YEAR = ComparatorUtil.chain(_BY_YEAR, _BY_TITLE);
-    private static final Comparator<Song> BY_YEAR_DESC = ComparatorUtil.chain(ComparatorUtil.reverse(_BY_YEAR), _BY_TITLE);
+    public static final Comparator<Song> BY_DISC_TRACK = ComparatorUtil.chain(_BY_DISC_TRACK, _BY_TITLE);
+    public static final Comparator<Song> BY_ALBUM = ComparatorUtil.chain(_BY_ALBUM, _BY_ALBUM_ID, BY_DISC_TRACK);
+    private static final Comparator<Song> BY_TITLE = ComparatorUtil.chain(_BY_TITLE, _BY_ARTIST, BY_ALBUM);
+    private static final Comparator<Song> BY_TITLE_DESC = ComparatorUtil.chain(ComparatorUtil.reverse(_BY_TITLE), _BY_ARTIST, BY_ALBUM);
+    private static final Comparator<Song> BY_ARTIST = ComparatorUtil.chain(_BY_ARTIST, _BY_TITLE, BY_ALBUM);
+    private static final Comparator<Song> BY_YEAR = ComparatorUtil.chain(_BY_YEAR, BY_TITLE);
+    private static final Comparator<Song> BY_YEAR_DESC = ComparatorUtil.chain(ComparatorUtil.reverse(_BY_YEAR), BY_TITLE);
 
     private static final List<SortOrder<Song>> SUPPORTED_ORDERS = Arrays.asList(
             Utils.build(
