@@ -40,6 +40,7 @@ import com.poupa.vinylmusicplayer.dialogs.LyricsDialog;
 import com.poupa.vinylmusicplayer.dialogs.SongShareDialog;
 import com.poupa.vinylmusicplayer.helper.MusicPlayerRemote;
 import com.poupa.vinylmusicplayer.helper.menu.SongMenuHelper;
+import com.poupa.vinylmusicplayer.misc.queue.IndexedSong;
 import com.poupa.vinylmusicplayer.model.Song;
 import com.poupa.vinylmusicplayer.model.lyrics.Lyrics;
 import com.poupa.vinylmusicplayer.ui.activities.base.AbsSlidingMusicPanelActivity;
@@ -177,6 +178,7 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
 
     @Override
     public void onMediaStoreChanged() {
+        // TODO If a song is removed from the MediaStore, this is not reflected in the playing queue untill restart
         updateQueue();
     }
 
@@ -198,7 +200,7 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
 
     @SuppressWarnings("ConstantConditions")
     private void updateCurrentSong() {
-        impl.updateCurrentSong(MusicPlayerRemote.getCurrentSong());
+        impl.updateCurrentSong(MusicPlayerRemote.getCurrentIndexedSong());
 
         // give the adapter a chance to update the decoration
         recyclerView.getAdapter().notifyDataSetChanged();
@@ -409,7 +411,7 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
     interface Impl {
         void init();
 
-        void updateCurrentSong(Song song);
+        void updateCurrentSong(IndexedSong song);
 
         void animateColorChange(final int newColor);
 
@@ -529,7 +531,7 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
         }
 
         @Override
-        public void updateCurrentSong(Song song) {
+        public void updateCurrentSong(IndexedSong song) {
             currentSong = song;
             currentSongViewHolder.title.setText(song.title);
             currentSongViewHolder.text.setText(MusicUtil.getSongInfoString(song));
@@ -574,7 +576,7 @@ public class CardPlayerFragment extends AbsPlayerFragment implements PlayerAlbum
         }
 
         @Override
-        public void updateCurrentSong(Song song) {
+        public void updateCurrentSong(IndexedSong song) {
             fragment.toolbar.setTitle(song.title);
             fragment.toolbar.setSubtitle(MusicUtil.getSongInfoString(song));
         }
