@@ -18,17 +18,7 @@ public class ComparatorUtil {
         };
     }
 
-    public static int compareLongInts(long v1, long v2) {
-        long diff = v1 - v2;
-
-        if (diff < Integer.MIN_VALUE) {
-            diff = Integer.MIN_VALUE;
-        }
-
-        if (diff > Integer.MAX_VALUE) {
-            diff = Integer.MAX_VALUE;
-        }
-
-        return (int)diff;
+    public static <E> Comparator<E> chain(Comparator<E> c1, Comparator<E> c2, Comparator<E> c3) {
+        return chain(c1, chain(c2, c3));
     }
 }
